requirements.txt
Django==1.11.2
django-anymail==0.10
django-cors-headers==2.1.0
django-filter==1.0.4
djangorestframework==3.6.3
Markdown==2.6.8
pytz==2017.2
redis==2.10.5
